Rename normalization to normalization_fwd (instances)
